% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFACSArrangedEgs.R
\name{plotFACSArrangedEgs}
\alias{plotFACSArrangedEgs}
\title{Arrange Example FACS Plots in stacks of Two}
\usage{
plotFACSArrangedEgs(
  egList,
  x,
  y,
  gate,
  ...,
  ncol = 1,
  xlim = c(-500, 20000),
  ylim = c(-500, 20000),
  biexTrans = "both"
)
}
\arguments{
\item{egList}{List of GatingHierarchies to plot in order given (so start with
WT!)}

\item{x}{String specifying x-axis parameter, as either channel or marker name}

\item{y}{String specifying y-axis parameter, as either channel or marker name}

\item{gate}{String specifying gate or gates (as character vector in that
case)}

\item{...}{Additional args to pass to plotFACSExample}

\item{ncol}{Numeric specifying the number of expected columns in the arranged
output}

\item{xlim}{Length-two numeric vector giving coord_cartesian the x-axis
limits}

\item{ylim}{Length-two numeric vector giving coord_cartesian the y-axis
limits}

\item{biexTrans}{String specifying whether a flowjo-style biex transform
should be applied to "both" (default), "x", or "y" axes. Any other input
results in no transform.}
}
\value{
A ggarrange'd collection of ggplot objects
}
\description{
This function creates the stacked example plots used throughout the
manuscript. Generally, these compare one WT and one IL23RKO example, but it
is extensible to other grids (such as the 2x2 used in the antibody graphs) by
simply passing a longer list of GatingHierarchies. By default, this also
applies biex scaling to both axes, and does NOT currently support modifying
their params from the default, so if that becomes important later I'll need
to add it in.
}
