% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFACSExample.R
\name{plotFACSExample}
\alias{plotFACSExample}
\title{Plot FACS data using pretty plotting defaults}
\usage{
plotFACSExample(
  data,
  x,
  y,
  gate,
  bins = 30,
  fontFix = TRUE,
  flowFix = TRUE,
  biex = "none",
  ...
)
}
\arguments{
\item{data}{A GatingHierarchy (for one plot) or GatingSet (for multiple).
This function will facet over the GatingSet as expected.}

\item{x}{A string specifying the parameter to plot on the X axis (either as
channel or marker name)}

\item{y}{A string specifying the parameter to plot on the Y axis. Currently,
this function does not support histograms, may update in the future.}

\item{gate}{A string specifying the gate (or vector of strings, for multiple
gates). Call gs_pop_get_paths on your GatingSet/Hierarchy for available
options.}

\item{bins}{A numeric specifying the number of bins to pass to
geom_density2d. Remember that this number should be WAY lower than what is
passed to geom_hex.}

\item{fontFix}{Boolean specifying whether the fonts should be shrunk down as
appropriate for figure publication-size plots. Defaults to TRUE so you
don't forget to do it!}

\item{flowFix}{Boolean specifying whether the FACS plots should be cleaned up
by removing the faceting strips and axis numbers. Defaults to TRUE so you
don't forget to do it!}

\item{biex}{A string specifying which axes ("x", "y", or "both") to apply a
default flowjo biex transform to. Leave it as "none" if you want no
trasform OR if you want to apply a non-default one manually after the fact
with scale_._flowjo_biex().}

\item{...}{Additional arguments to pass to geom_stats. "adjust", "position",
"size", and "type" are all useful here. See stat_position for info about
adjust and position.}
}
\value{
A ggplot object containing the plotted FACS data
}
\description{
This is meant to produce a publication-quality FACS plot, but can also be
used for exploratory analysis. Note that this function expects to draw GATED
data - you'll need to run a raw ggcyto call (or make a new function for
ungated) if you don't have a gate in mind.
}
